requirements.txt
# --- Runtime ---
a2a-sdk[http-server]>=0.2.0
sse-starlette>=2.0.0
google-genai>=0.3.0
httpx>=0.27.0
google-search-results>=2.4.2
trafilatura>=1.12.0
pyyaml>=6.0.1
pydantic>=2.7.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
starlette>=0.37.0

# --- Dev / test ---
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
respx>=0.21.0
freezegun>=1.5.0
